run: Option to override default 1 day timeout
Tuxmake spwans all of its containers with the command `sleep 1d`. This
is useful as it means that if the container becomes dangling somehow, it
automatically shuts down after 1 day. This timeout is sufficient for all
imaginable build configs. However there are cases where we want to run
the FVP for longer than this; the glibc test suite takes 3-4 days to
complete.

So let's add a command line option, --timeout, to allow the user to
specify the number of days after which the container should auto
shutdown. When not provided, we stick with tuxmake's default.

Unfortunately, tuxmake has no API to set this, so we achieve it by
hooking its spawn_container() method and modifying the argument. Yuk;
this is fragile.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>