test: Move cca-[34]world.yaml to test.py's CONFIGS table
cca-3world.yaml and cca-4world.yaml have traditionally been special
cased by the test framework because they have their arch revision baked
into the config and don't support overlaying different arches.

However, this is totally unecessary now that we expose which arches are
supported in the CONFIGS table. So let's extend that so that not
specifiying any arches means "don't add an arch overlay". Then we can
add them to the table with everything else.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>