test: Run test-sync.sh repo sync tests on CI
Jean-Philippe's test-sync.sh does an excellent job of testing
Shrinkwrap's various repo sync options work as intended. So let's hook
it up so the CI runs it automatically as a regression test.

For simplicity, let's call it from within test.py so that we can easily
add a junit test result for it, which Gitlab will display.

Unfortunately there are some complications for running in the container;
We can't use ssh for the remote submodules, so let's use file://, but
this requires overriding a global config to allow it. And we must
configure name and email. So let's create a temporary gitconfig for the
test.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>