run: Update FVP version to 11.27
As of 11.27, ScalableVectorExtension.so plugin is no longer included
with the model. Instead, the equivalent functionality is included
directly in the model so remove references to the plugin.

SVE/SME configuration parameters have also moved, so we also update the
relevant configs:

SVE.ScalableVectorExtension.<parameter> -> cluster[01].sve.<parameter>
SVE is enabled via cluster[01].has_sve=1

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>