docs: Remove refs to ScalableVectorExtension.so
Since the ScalableVectorExtension.so plugin functionality is now
implemented directly by the FVP, let's remove references to it and it's
parameters from the documentation.

It turns out some of the output examples in the quickstart guide were a
little out of date, hence the noise introduced by generating new ones.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>