docker: Don't add FVP plugins directory to PATH
Modern FVPs (sometime between 11.24 and 11.27) no longer mark its
plugins as being executable. This means we can no longer rely on the
convention of adding the plugins directory to the PATH then locating a
plugin with `which GenericTrace.so`.

Given this no longer works, let's simplify by removing the code that set
it up.

Fortunately the only usage of this trick within bundled configs was for
ScalableVectorExtension.so. But that's just been removed so there are no
more users relying on it.

If any out-of-tree configs require locating a bundled plugin, the
recommended hack is:

$$(dirname $$(which FVP_Base_RevC-2xAEMvA))/../../plugins/$$(basename $$(dirname $$(which FVP_Base_RevC-2xAEMvA)))/<plugin>.so

This will cause bash to construct a path relative to the FVP model. The
above trick should work for both x86_64 and aarch64.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>