ci: Workaround issue in Kaniko
After updating to FVP 11.27, wget fails when trying to download it from
developer.arm.com from within the Kaniko container on the CI, when
running on an amd64 system:

  wget: note: TLS certificate validation not implemented
  wget: TLS error from peer (alert code 80): internal error
  wget: error getting response: Connection reset by peer

This is caused by a busybox/musl issue introduced in v1.9.2 of the
Kaniko container image. See
https://github.com/GoogleContainerTools/kaniko/issues/2452.

Work around this by pinning to v1.9.1 as suggested in the issue.

However, while the tip works for arm64, v1.9.1 causes a different
problem:

  Connecting to developer.arm.com (184.28.198.219:443)
  wget: error getting response: Connection reset by peer

So use 'debug' tag for arm64 and 'v1.9.1-debug' tag for amd64. Yuk.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>