config: Update edk2 revision to workaround "subhook" issue
Checking out edk2 has started failing due to one of it's submodule
dependencies, https://github.com/Zeex/subhook.git, being removed from
GitHub:

  Cloning into '/data_nvme0n1/ryarob01/shrinkwrap_demo/workspace/edk2/edk2/UnitTestFrameworkPkg/Library/SubhookLib/subhook'...
  remote: Repository not found.
  fatal: repository 'https://github.com/Zeex/subhook.git/' not found
  fatal: clone of 'https://github.com/Zeex/subhook.git' into submodule path '/data_nvme0n1/ryarob01/shrinkwrap_demo/workspace/edk2/edk2/UnitTestFrameworkPkg/Library/SubhookLib/subhook' failed
  Failed to clone 'UnitTestFrameworkPkg/Library/SubhookLib/subhook'. Retry scheduled

This has been resolved in edk2's master branch by linking to a mirror of
the project. But that commit is not yet in any stable tag. So for now,
let's fix to the SHA of the fix. Once the next stable tag is out, we
will need to update again.

However, updating to this SHA breaks the build because edk2-platforms is
out of date. So let's update it to its latest commit also. But this is
broken for RELEASE builds, so we must build DEBUG. A patch will be
merged soon to fix RELEASE, so we will update to that once merged.

Finally, the downstream branch of edk2 used for the CCA configs (both
host and guest) need to be updated. A new branch has been created that
cherry-picks the submodule fix.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>