test: Move assets.yaml to a private config store
assets.yaml defines the test components that need to be built for use by
test.py. However, it was previously part of the built-in config store so
was public to all users. Let's move it to a config store in the test
directory and set SHRINKWRAP_CONFIG to point to that store. This makes
it private to genassets.sh.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>