config: Split edk2 and edk2-platforms into independent targets
Allow each to be separately configured with local sourcedir's
as this is currently not supported by the nested repo field syntax

This introduces a new issue as the EDK2 build system will always
search for edk2-platforms in the base repo ("sourcedir"), ignoring
the value of PACKAGES_PATH, unless WORKSPACE points elsewhere.

Take this opportunity to override the default build path in a more
widely compatible way, since EDK2_OUT_DIR is not universally
supported. Same with CONF_PATH so that cleaning the build becomes
as trivial as deleting "builddir".

Signed-off-by: Yuliang Wang <yuliang.wang@arm.com>