configs: Update kvm-unit-tests
Update kvm-unit-tests and make it easier to use the final product.

Rather than archiving the kvm-unit-tests directory, share it as is with
the guest. I've had issues uncompressing the archive from inside the
guest, because the buildroot tar didn't recognize the format:

    tar: invalid tar magic

Update the generated config.mak so we can run the test script from
within the guest.

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>