build, run: Use host network to avoid VPN issues
There have been numerous reports of VPNs not playing nicely with the
bridged docker network. Let's switch to using the host network to try to
solve this. The host network is described as "Remove network isolation
between the container and the Docker host". Essentially they are now the
same network entity.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>