docker: install uv in Shrinkwrap's Docker images
This small change ensures that [uv][1] is available and usable from
within Shrinkwrap's Docker images, and uses it to set up a new global
virtual environment.

For the moment, the virtual environment that is already created with
venv during the build process continues to take precedence - this change
is just about ensuring that uv is available and functional.

[1]: https://docs.astral.sh/uv/
[2]: https://docs.python.org/3/library/venv.html