docker: include entire repository in build context
This change makes the entire repository available from the Docker build
context, which means that we can access files outside of the `docker`
directory from the image building process.

The reason for this is so that we can install Python runtime
dependencies based on the `pyproject.toml` file in the root of the
repositroy, rather than having to manage them manually in the
Dockerfile.

Signed-off-by: Chris Kay <chris.kay@arm.com>