project: suppress exceptions by default in package context
When calling Shrinkwrap from a script context, exceptions triggered by
the program are suppressed by default. This does not occur when calling
the `main()` function directly.

The changes that follow this one introduce a project script which calls
the `main()` function, meaning it does not share this behaviour.

This change moves this exception suppression behaviour out of the global
scope and into the function so that the observed behaviour in both a
script context and a package context is identical.

Signed-off-by: Chris Kay <chris.kay@arm.com>