run: Don't show update_submodules() bash function
Commit 22e540d ("submodules: Support for --reference when cloning
submodules") added a bash update_submodules() script to
script_preamble(). The script is used when building to checkout the
repository for a component. But script_preamble() provides the preamble
for all commands (build, clean, run). It is not needed for clean and
run, so let's hide it from those.

This tidies up the --dry-run output.

Fixes: 22e540d ("submodules: Support for --reference when cloning submodules")
Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>