ci: Ensure caches get correctly invalidated
Previously if .cache-exists was found in the docker-assets cache, no
attempt was made to download the required files into the cache. This is
"safe" since when building the container, if the required files are not
present, they are downloaded. But this is problematic for 2 reasons; it
is slow. The whole point of the cache is to avoid the download. But more
importantly, the kaniko image that we are stuck using to create the
image provides a buggy wget so it fails.

build.sh, when invoked with --version none, will already avoid download
if the correct files are present, so let's just always call that to
(potentially) repopulate the cache. Then kaniko's wget never needs to be
used.

While we are at it, let's rework to always delete anything in the cache
if requested by the user.

Fixes: ec08849 ("ci: Add gitlab pipeline definition")
Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>