config: Workaround bootwrapper's lack of support for GCS
bootwrapper doesn't disable the GCS EL3 traps so the kernel traps to EL3
when trying to configure GCS if present. As a temporary workaround until
this is fixed in bootwrapper, let's avoid testing configs with
bootwrapper where GCS is enabled.

This should be reverted when bootwrapper gains support. Tracked in
PDSWLINUX-4668.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>