test: Honour config-defined container image
Commit a502ce3 ("feature: add an optional 'image' field to config
files.") made it possible for a config to specify it's preferred
container image. When used with buildall, any configs that specify an
image must specify the same image or the build fails. But test.py was
just always running with the base-full image. This works because
currently all configs that have special requirements, require base-full
and base-full is a superset of base-slim.

But this may change in future meaning test.py would break. Additionally,
it would be nice for test.py to use the corresponding tag for the
release instead of just using "latest", which could introduce some
compat issues.

So let's make a change so that the batching takes into account the
preferred container image.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>