containers: tag containers with shrinkwrap version
The previous strategy of publishing the containers under the "latest"
tag meant that we were forcing users to always use the bleeding edge.
Let's transition to a better tagging strategy by additionally tagging
container images with the current shrinkwrap version. This means that as
long as we systematically rev the version number, we can end up with
locked-down and reproducible pairings. Additionally shrinkwrap will now
default to using it's version as the image tag when a tag is not
provided.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>