version: Switch to new versioning scheme: 2025.12.0.dev0
We are introducing a new versioning scheme as follows:

<year>.<month>.<patch>[.dev0]

The year and the month are the target for the next release. Patch is
usually 0, but can be incremented if we need to make a point release for
bug fixing. .dev0 indicates the release is under development.

So as a concrete example, we are aiming to make the first release under
these new rules in December 2025. The version will be 2025.12.0. While
we are in the development cycle, the version is set to 2025.12.0.dev0.
At the end of the cycle we will remove the .dev0 and tag the release.
Then we will immediately set the version for the next version (e.g.
March): 2026.3.0.dev0.

This approach makes it easy to tell when a release was made and also
conforms to Python versioning requirements.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>