ci: Don't run pipeline for tag or merge request events
We are already running the pipeline whenever a branch is pushed so these
extra triggers are unnecessary. Additionally, the readthedocs build
script only works when there is a branch name that it can convert to an
RTD version. For these 2 triggers no such branch exists. So this
approach works around that issue.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>