refactor: use isinstance() instead of type()
Use of isinstance() is canonical.

isinstance() will match on the specified type and its subclasses, whereas comparing the
result of type() is an exact class match which ignores inheritance (the compare should be
done with is/is not). Unless there is a specific reason to require an exact match,
isinstance() should be used to avoid future maintenance issues. The uses in shrinkwrap do
not require exact matches, so isinstance() should be used instead.

Signed-off-by: Chris Reed <chris.reed@arm.com>