test: Bypass attestation checks when built on arm64
test.py contains a config that boots a realm guest inside the host then
in the guest, attempts attestation using the cca-workload-attestation
tool. However, when this test runs on an arm64 machine, the tool is not
present in the buildroot image. This causes the guest not to poweroff.
Which means we hit the test timeout and the test fails.

I'm not sure exactly why the tool is not present, but [1] shows that it
depends on some properties of the build host:

  | config BR2_PACKAGE_CCA_WORKLOAD_ATTESTATION
  |     bool "CCA Workload Attestation PoC"
  |     depends on BR2_PACKAGE_HOST_GO_TARGET_ARCH_SUPPORTS
  |     depends on BR2_PACKAGE_HOST_GO_TARGET_CGO_LINKING_SUPPORTS

Perhaps arm64 does not satisfy these requirements? That would explain
the issue at least because this component would get disabled and not
built into the buildroot image.

This all needs further investigation. So for now, let's bypass doing the
attestation check if the tool is not present. This will at least mean we
continue to get full coverage for x86 build hosts and will still allow
testing that booting a guest in a realm works for arm64 build hosts.

[1] https://git.codelinaro.org/linaro/dcap/buildroot-external-cca/-/blob/main/package/cca-workload-attestation/Config.in?ref_type=heads

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>