test: Fix test names to include overlays
Commit a26498b ("test: Gather tests by overlay") renamed the "overlay"
field in the test result structure to "overlays" ands started storing
them as a comma-separated list. But the test name generator function was
not updated. As a result it was failing to add the overlays to the test
name and therefore many tests had the same name. This confused gitlab,
which collapsed all the tests with the same name into a single test in
it's UI.

Fixes: a26498b ("test: Gather tests by overlay")
Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>